backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
SQLAlchemy
psycopg2-binary
qdrant-client
sentence-transformers
numpy
scikit-learn
langdetect
PyMuPDF
python-docx
pillow
pytesseract
python-multipart
requests
python-dotenv
tenacity
google-generativeai
